Made a utils folder in app, server, and common for functions for front end, back end, and both respectively. Moved functions into appropriate folders and fixed code broken by the moves.
