+++ app/utils/functions.js
@@ -1,0 +1,2 @@
+// frontend only functions go here
+module.exports = {};
