+++ server/reducer.js
@@ -8,7 +8,8 @@ import {
   ADD_SONG_FROM_HISTORY
 } from '../common/constants/ActionTypes';
 
-const https = require('https');
+const commonFunctions = require('../common/utils/functions.js');
+const api = require('./utils/APIcalls.js');
 const moment = require('moment');
 
 export const initialState = {
@@ -17,75 +18,6 @@ export const initialState = {
 };
 
 // We should probably move all of this somewhere else
-function getVidFromUrl(url) {
-  // lazy query string parse for vid
-  if (url.indexOf('v=') === -1) {
-    return '';
-  }
-  const temp = url.split('v=');
-  const vid = temp[1].split('&');
-  return vid[0];
-}
-
-function updateSong(url) {
-  const song = {};
-  song.url = url;
-  song.vid = getVidFromUrl(url);
-
-  if (song.vid === '') {
-    console.log('Invalid URL. Not adding song.');
-    return song;
-  }
-
-  song.artist = null;
-  song.duration = null;
-  song.src = null;
-  song.title = null;
-  song.uploadDate = null;
-  song.upvotes = 0;
-  song.userUpvotes = [];
-  song.thumbnail = null;
-  song.endedAt = null;
-
-  // key has to be passed in as an environment varibale
-  // Example: YOUTUBE_API=aksdfjalksdfjalskdfjlk npm start
-  if (!process.env.YOUTUBE_API) {
-    console.log('You didnt put in an API key correctly so partyq will not pull song information from the YouTube API.');
-    console.log('To run with API key do $YOUTUBE_API={API_KEY_HERE} npm start');
-    return song;
-  }
-
-  const callAPIURL = 'https://www.googleapis.com/youtube/v3/videos?part=snippet%2C+contentDetails&id='
-                 + song.vid + '&key=' + process.env.YOUTUBE_API;
-
-  https.get(callAPIURL, function(res) {
-    let data = '';
-    res.on('data', function(chunk) {
-      data += chunk;
-    });
-    res.on('end', function() {
-      const youTubeSongData = JSON.parse(data);
-      if (youTubeSongData.error) {
-        console.log('API error');
-        console.log('Reason: ' + youTubeSongData.error.errors[0].reason);
-        console.log('Message: ' + youTubeSongData.error.errors[0].message);
-      } else if (!youTubeSongData.items[0]) {
-        console.log('Invalid VID: ' + song.vid);
-      } else {
-        song.duration = moment.duration(youTubeSongData.items[0].contentDetails.duration).asMilliseconds();
-        song.artist = youTubeSongData.items[0].snippet.channelTitle;
-        song.duration = moment.utc(song.duration).format('mm:ss');
-        song.title = youTubeSongData.items[0].snippet.title;
-        song.uploadDate = youTubeSongData.items[0].snippet.publishedAt;
-        song.thumbnail = youTubeSongData.items[0].snippet.thumbnails.default.url;
-        song.src = 'youtube';
-        console.log(song);
-      }
-      return song;
-    });
-  });
-  return song;
-}
 
 function queueReducer(state = initialState.queue, action) {
   const queueSonglist = state.songlist;
@@ -93,7 +25,10 @@ function queueReducer(state = initialState.queue, action) {
   const userid = action.id;
   switch (action.type) {
   case ADD_SONG:
-    const song = updateSong(action.url);
+    if (!commonFunctions.linkIsValid(action.url)) {
+      return state;
+    }
+    const song = api.callYoutube(action.url, commonFunctions.getVidFromUrl(action.url));
     if (song.vid === '') {
       return state;
     }
